requirements.txt
pytest-asyncio==0.15.1
aiomysql==0.0.21
requests==2.25.1
aiohttp==3.7.4.post0
